pyproject.toml
[tool.poetry]
name = "cbbd_etl"
version = "0.1.0"
description = "CollegeBasketballData ETL pipeline"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cbbd_etl", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
boto3 = "^1.34.0"
pyarrow = "^16.0.0"
pyyaml = "^6.0.1"
python-dateutil = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
